+++ Cyotek.Windows.Forms.ImageBox.Demo/MainForm.Designer.cs
@@ -58,18 +58,20 @@
       this.selectAllToolStripButton = new System.Windows.Forms.ToolStripButton();
       this.selectNoneToolStripButton = new System.Windows.Forms.ToolStripButton();
       this.zoomLevelsToolStripComboBox = new System.Windows.Forms.ToolStripComboBox();
+      this.toolStripSeparator4 = new System.Windows.Forms.ToolStripSeparator();
+      this.virtualModeToolStripButton = new System.Windows.Forms.ToolStripButton();
       this.menuStrip = new System.Windows.Forms.MenuStrip();
       this.fileToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
       this.openToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
       this.toolStripSeparator3 = new System.Windows.Forms.ToolStripSeparator();
       this.exitToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
+      this.helpToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
+      this.aboutToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
       this.editToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
       this.copyToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
       this.toolStripSeparator6 = new System.Windows.Forms.ToolStripSeparator();
       this.selectAllToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
       this.selectNoneToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
-      this.helpToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
-      this.aboutToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
       this.viewSplitContainer.Panel1.SuspendLayout();
       this.viewSplitContainer.Panel2.SuspendLayout();
       this.viewSplitContainer.SuspendLayout();
@@ -95,8 +97,9 @@
       // viewSplitContainer.Panel2
       // 
       this.viewSplitContainer.Panel2.Controls.Add(this.propertyGrid);
-      this.viewSplitContainer.Size = new System.Drawing.Size(909, 455);
-      this.viewSplitContainer.SplitterDistance = 609;
+      this.viewSplitContainer.Size = new System.Drawing.Size(1060, 536);
+      this.viewSplitContainer.SplitterDistance = 759;
+      this.viewSplitContainer.SplitterWidth = 5;
       this.viewSplitContainer.TabIndex = 0;
       // 
       // selectionSplitContainer
@@ -114,8 +117,9 @@
       // selectionSplitContainer.Panel2
       // 
       this.selectionSplitContainer.Panel2.Controls.Add(this.previewImageBox);
-      this.selectionSplitContainer.Size = new System.Drawing.Size(609, 455);
-      this.selectionSplitContainer.SplitterDistance = 305;
+      this.selectionSplitContainer.Size = new System.Drawing.Size(759, 536);
+      this.selectionSplitContainer.SplitterDistance = 385;
+      this.selectionSplitContainer.SplitterWidth = 5;
       this.selectionSplitContainer.TabIndex = 1;
       // 
       // imageBox
@@ -124,7 +128,7 @@
       this.imageBox.Dock = System.Windows.Forms.DockStyle.Fill;
       this.imageBox.Location = new System.Drawing.Point(0, 0);
       this.imageBox.Name = "imageBox";
-      this.imageBox.Size = new System.Drawing.Size(609, 305);
+      this.imageBox.Size = new System.Drawing.Size(759, 385);
       this.imageBox.TabIndex = 0;
       this.imageBox.Selected += new System.EventHandler<System.EventArgs>(this.imageBox_Selected);
       this.imageBox.SelectionRegionChanged += new System.EventHandler(this.imageBox_SelectionRegionChanged);
@@ -147,7 +151,7 @@
       this.previewImageBox.ImageBorderStyle = Cyotek.Windows.Forms.ImageBoxBorderStyle.FixedSingle;
       this.previewImageBox.Location = new System.Drawing.Point(0, 0);
       this.previewImageBox.Name = "previewImageBox";
-      this.previewImageBox.Size = new System.Drawing.Size(609, 146);
+      this.previewImageBox.Size = new System.Drawing.Size(759, 146);
       this.previewImageBox.TabIndex = 0;
       // 
       // propertyGrid
@@ -158,7 +162,7 @@
       this.propertyGrid.Location = new System.Drawing.Point(0, 0);
       this.propertyGrid.Name = "propertyGrid";
       this.propertyGrid.SelectedObject = this.imageBox;
-      this.propertyGrid.Size = new System.Drawing.Size(296, 455);
+      this.propertyGrid.Size = new System.Drawing.Size(296, 536);
       this.propertyGrid.TabIndex = 0;
       // 
       // statusStrip
@@ -170,16 +174,17 @@
             this.zoomToolStripStatusLabel,
             this.cursorToolStripStatusLabel,
             this.selectionToolStripStatusLabel});
-      this.statusStrip.Location = new System.Drawing.Point(0, 504);
+      this.statusStrip.Location = new System.Drawing.Point(0, 585);
       this.statusStrip.Name = "statusStrip";
+      this.statusStrip.Padding = new System.Windows.Forms.Padding(1, 0, 16, 0);
       this.statusStrip.ShowItemToolTips = true;
-      this.statusStrip.Size = new System.Drawing.Size(909, 22);
+      this.statusStrip.Size = new System.Drawing.Size(1060, 22);
       this.statusStrip.TabIndex = 1;
       // 
       // toolStripStatusLabel1
       // 
       this.toolStripStatusLabel1.Name = "toolStripStatusLabel1";
-      this.toolStripStatusLabel1.Size = new System.Drawing.Size(814, 17);
+      this.toolStripStatusLabel1.Size = new System.Drawing.Size(963, 17);
       this.toolStripStatusLabel1.Spring = true;
       // 
       // autoScrollPositionToolStripStatusLabel
@@ -230,10 +235,12 @@
             this.toolStripSeparator2,
             this.selectAllToolStripButton,
             this.selectNoneToolStripButton,
-            this.zoomLevelsToolStripComboBox});
+            this.zoomLevelsToolStripComboBox,
+            this.toolStripSeparator4,
+            this.virtualModeToolStripButton});
       this.toolStrip.Location = new System.Drawing.Point(0, 24);
       this.toolStrip.Name = "toolStrip";
-      this.toolStrip.Size = new System.Drawing.Size(909, 25);
+      this.toolStrip.Size = new System.Drawing.Size(1060, 25);
       this.toolStrip.TabIndex = 2;
       // 
       // openToolStripButton
@@ -337,8 +344,23 @@
       // 
       this.zoomLevelsToolStripComboBox.DropDownStyle = System.Windows.Forms.ComboBoxStyle.DropDownList;
       this.zoomLevelsToolStripComboBox.Name = "zoomLevelsToolStripComboBox";
-      this.zoomLevelsToolStripComboBox.Size = new System.Drawing.Size(121, 25);
+      this.zoomLevelsToolStripComboBox.Size = new System.Drawing.Size(140, 25);
       this.zoomLevelsToolStripComboBox.SelectedIndexChanged += new System.EventHandler(this.zoomLevelsToolStripComboBox_SelectedIndexChanged);
+      // 
+      // toolStripSeparator4
+      // 
+      this.toolStripSeparator4.Name = "toolStripSeparator4";
+      this.toolStripSeparator4.Size = new System.Drawing.Size(6, 25);
+      // 
+      // virtualModeToolStripButton
+      // 
+      this.virtualModeToolStripButton.DisplayStyle = System.Windows.Forms.ToolStripItemDisplayStyle.Image;
+      this.virtualModeToolStripButton.Image = global::Cyotek.Windows.Forms.Demo.Properties.Resources.VirtualMode;
+      this.virtualModeToolStripButton.ImageTransparentColor = System.Drawing.Color.Magenta;
+      this.virtualModeToolStripButton.Name = "virtualModeToolStripButton";
+      this.virtualModeToolStripButton.Size = new System.Drawing.Size(23, 22);
+      this.virtualModeToolStripButton.Text = "Virtual Mode";
+      this.virtualModeToolStripButton.Click += new System.EventHandler(this.virtualModeToolStripButton_Click);
       // 
       // menuStrip
       // 
@@ -348,7 +370,8 @@
             this.helpToolStripMenuItem});
       this.menuStrip.Location = new System.Drawing.Point(0, 0);
       this.menuStrip.Name = "menuStrip";
-      this.menuStrip.Size = new System.Drawing.Size(909, 24);
+      this.menuStrip.Padding = new System.Windows.Forms.Padding(7, 2, 0, 2);
+      this.menuStrip.Size = new System.Drawing.Size(1060, 24);
       this.menuStrip.TabIndex = 3;
       // 
       // fileToolStripMenuItem
@@ -383,6 +406,21 @@
       this.exitToolStripMenuItem.Size = new System.Drawing.Size(146, 22);
       this.exitToolStripMenuItem.Text = "E&xit";
       this.exitToolStripMenuItem.Click += new System.EventHandler(this.exitToolStripMenuItem_Click);
+      // 
+      // helpToolStripMenuItem
+      // 
+      this.helpToolStripMenuItem.DropDownItems.AddRange(new System.Windows.Forms.ToolStripItem[] {
+            this.aboutToolStripMenuItem});
+      this.helpToolStripMenuItem.Name = "helpToolStripMenuItem";
+      this.helpToolStripMenuItem.Size = new System.Drawing.Size(44, 20);
+      this.helpToolStripMenuItem.Text = "&Help";
+      // 
+      // aboutToolStripMenuItem
+      // 
+      this.aboutToolStripMenuItem.Name = "aboutToolStripMenuItem";
+      this.aboutToolStripMenuItem.Size = new System.Drawing.Size(152, 22);
+      this.aboutToolStripMenuItem.Text = "&About...";
+      this.aboutToolStripMenuItem.Click += new System.EventHandler(this.aboutToolStripMenuItem_Click);
       // 
       // editToolStripMenuItem
       // 
@@ -428,30 +466,16 @@
       this.selectNoneToolStripMenuItem.Text = "Select &None";
       this.selectNoneToolStripMenuItem.Click += new System.EventHandler(this.selectNoneToolStripMenuItem_Click);
       // 
-      // helpToolStripMenuItem
-      // 
-      this.helpToolStripMenuItem.DropDownItems.AddRange(new System.Windows.Forms.ToolStripItem[] {
-            this.aboutToolStripMenuItem});
-      this.helpToolStripMenuItem.Name = "helpToolStripMenuItem";
-      this.helpToolStripMenuItem.Size = new System.Drawing.Size(44, 20);
-      this.helpToolStripMenuItem.Text = "&Help";
-      // 
-      // aboutToolStripMenuItem
-      // 
-      this.aboutToolStripMenuItem.Name = "aboutToolStripMenuItem";
-      this.aboutToolStripMenuItem.Size = new System.Drawing.Size(116, 22);
-      this.aboutToolStripMenuItem.Text = "&About...";
-      this.aboutToolStripMenuItem.Click += new System.EventHandler(this.aboutToolStripMenuItem_Click);
-      // 
       // MainForm
       // 
-      this.AutoScaleDimensions = new System.Drawing.SizeF(6F, 13F);
+      this.AutoScaleDimensions = new System.Drawing.SizeF(7F, 15F);
       this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
-      this.ClientSize = new System.Drawing.Size(909, 526);
+      this.ClientSize = new System.Drawing.Size(1060, 607);
       this.Controls.Add(this.viewSplitContainer);
       this.Controls.Add(this.statusStrip);
       this.Controls.Add(this.toolStrip);
       this.Controls.Add(this.menuStrip);
+      this.Font = new System.Drawing.Font("Segoe UI", 9F);
       this.Icon = ((System.Drawing.Icon)(resources.GetObject("$this.Icon")));
       this.MainMenuStrip = this.menuStrip;
       this.Name = "MainForm";
@@ -513,6 +537,8 @@
     private System.Windows.Forms.ToolStripMenuItem aboutToolStripMenuItem;
     private System.Windows.Forms.ToolStripComboBox zoomLevelsToolStripComboBox;
     private System.Windows.Forms.ToolStripMenuItem selectNoneToolStripMenuItem;
+    private System.Windows.Forms.ToolStripSeparator toolStripSeparator4;
+    private System.Windows.Forms.ToolStripButton virtualModeToolStripButton;
   }
 }